requirements.txt
requests
pywebview

